refactor: 게시물 삭제 api 연결
